kafka-e2e-test-tool/old_version/requirements.txt
# Core dependencies
click>=8.0.0
PyYAML>=6.0

# Kafka (optional - will use mock if not available)
kafka-python>=2.0.2

# Validation (optional)
jsonschema>=4.0.0

# Testing with containers (optional)
testcontainers>=3.4.0
